requirements.txt
# Building and Demonstrating Map Servers with OpenAI Agents SDK
# Part 2: Implementation - Dependencies

# OpenAI Agents SDK
openai>=1.50.0

# Async HTTP client
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: for better CLI experience
rich>=13.0.0
